Create Task2
This is a Python-based Tic Tac Toe game made as part of my CodSoft Internship (Task 2). The game allows a human player (X) to play against a computer AI (O). The AI is designed using the Minimax algorithm, which makes it unbeatable – it will either win or force a draw. The project helped me learn about game logic, AI decision-making, and Python programming.